Telephone Network
Dijkstra's algorithm is a shortest-path algorithm. 
Here, I have tried to change it into a longest-path algorithm, using the example of a telephone network.

The telephone network is the graph, the cell towers are the nodes of the graph, the edges are the routing paths, and the edge weights are the bandwidths.
While routing a call, the greater the bandwidth, more the information that can be transmitted. With this in mind, the code has been written to find the longest distance between any two pairs of nodes.

The code does not always work, because of self-loop issues, and because a solution cannot be found in polynomial time.